requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch==2.0.1+cu117
torchvision==0.15.2+cu117
torchaudio==2.0.2+cu117
transformers==4.33.2
tqdm==4.66.1
pandas==2.1.1
scikit-learn==1.3.0
numpy<2.0
openpyxl==3.1.2
openai-whisper==20240927
streamlit==1.25.0
sounddevice==0.4.6
pygame==2.3.0
nltk==3.8.1
librosa==0.10.0.post2
